feat(operator): Add a new dragonflydb operator
This commit adds a new Kubernetes operator written in Golang into
the repo, that can create and manage DragonFly Instances for Users.
The current main reason for using this operator over an Helm chart
is for the failover capabilities, as it can automatically mark
a replica as a master incase of a failure.